Expose standalone metric for number of messages per queue
This commit adds a new gauge, "showq_message_count", containing the
total number of messages by queue.

It was already sort of possible to get this information by piggybacking
off of the message size and/or message age histograms, however the
downside of that approach is that when there are no messages in a given
queue, no metric is exported.

This new showq_message_count metric will instead export 0 when a given
queue contains no messages, allowing one to query instances where a
postfix queue is empty and preventing metrics from being considered
stale by Prometheus.